requirements.txt
flask>=2.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
numpy>=1.21.0
Pillow>=9.0.0
python-doctr>=0.7.0
pdf2image>=1.16.3
paddleocr>=2.6.0